Linear search (2.1-3): implementation + tests
Another very simple algorithm, this time from one of the exercises. This
time, most of the code actually goes into tests, since this commit adds
some testing infrastructure in the new `testing::vector` module, to
enable us to efficiently write quickcheck cases in the actual test files.

Some documentation is also included, and of course we bump the version
by 1.