Merge sort (2.3): implementation + tests
Our first "good" sorting algorithm! The implementation works, but I'm
not happy with it and probably going to rework it. Reading through it, I
realized that I was being a huge doofus and made several unwise
decisions which make the code smack of 90s era C.

The tests are the usual lot for sorting algorithms.