Binary Search (2.3-5): implementation + tests
Our first and only (?) non-trivial search algorithm, binary search. The
implementation is straightforward, and tests are included as usual.

I also realized while writing this that linear search does not return
the index of the found element, but the element itself, because, uhh,
I... forgot what I was doing? Expect a fix in the very near future.