Linear Search (2.1-3): correction/rewrite
As alluded to in the last commit, the implementation of linear search we
had was weird, in that it didn't return the index of the found element,
but the element itself. As it turns out, I had a good reason for writing
it like that: it's the only version of the algorithm that works on
iterators. However, to keep in the spirit of the book, I think it would
be wise to limit ourselves to arrays (vectors) and return the found
index, so this commit updates the algorithm, its documentation, and the
tests.