Shamo's Algorithm (4.1): implementation + tests
This algorithm isn't officially named in the book, so we give it the
name of its inventor simply to be able to distinguish it from other
algorithms solving the same task, which is the maximum subarray problem.
This is one of the longer algorithms implemented so far, not necessarily
because it does much more than other algorithms, but simply because
divide-and-conquer always seems to turn out pretty wordy.

The tests are very lacklustre. I'll improve them later, once I've
implemented at least one other algorithm solving the same problem.