Max. Subarr. Algs.: improve tests
As promised, here's an expansion on correctness testing for the maximum
subarray algorithms, in the form of agreement tests in the new
`maximum_subarray::test` module. Basically, instead of checking each
algorithm for bugs, we just check them against each other, so that
either both are correct or both are equally wrong (for the same input).

To make these tests somewhat less tedious to write, we include some more
vector testing utilities, namely for generating arbitrary vectors of
strictly positive and strictly negative elements.